requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
django-environ>=0.11.0
django-cors-headers>=4.0.0
django-simple-history>=3.3.0
django-encrypted-model-fields>=0.6.5
nltk>=3.8.0
textblob>=0.17.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0
Pillow>=10.0.0
gunicorn>=21.0.0
python-decouple>=3.8
